EDA and Best Subsets
Hello!  I created a subset of the selected variables, created training and test sets, ran a base MLR with all variables for comparison, and then ran best subsets with cross-validation.  I think you guys may need to change the location of the data file when you run the code, but it should all work!